Welcome to Hell! Check Description.
Start by taking a look at the toDo.txt file I made. Then take a look at the code I have already. We Will most likely need to re-write a lot of it so if you can't figure out a way to make it work easily just delete it and start over.